fix the codeclimate issues
