+++ src/controllers/statistics.js
@@ -1,8 +1,8 @@
 import models from '../models'
-import UserServices from '../services/user.service';
+import top5VisitedLocation from '../helper/top5visitedloc'
+import top5BookedAccommodation from '../helper/top5VisitedBookedAcc'
 
 const getStatistics = async function (req, res, next) {
-   
     const { Op } = require("sequelize");
     const activeUsers = await models.User.findAll({
         where: {
@@ -11,50 +11,14 @@ const getStatistics = async function (req, res, next) {
           }
         }
         })
+        //Get the number of  Accomodation
     const numberOfAccommodation = await models.Accommodation.findAndCountAll({})
+    //Get the number of Locations
     const numberOfLocation = await models.Location.findAndCountAll({})
-    const locations = await models.Location.findAll({
-        attributes: ['LocationName']
-    })
-    let locationArr = {}
-    locations.forEach(location => {
-        const name = location.LocationName
-        if(locationArr.hasOwnProperty(name)){
-            locationArr[name] = locationArr[name] + 1
-        } else{
-            locationArr[name] = 1
-        }
-    });
-    let sortedLocations = []
-    
-    for( const key in locationArr){
-        sortedLocations.push([key,locationArr[key]])
-    }
-
-    sortedLocations.sort(function(a, b) {
-        return b[1] - a[1];
-    });
-  
-const accommodations = await models.Booking.findAll({
-    attributes: ['accommodationId','id']
-})
-let accommodationArr = {}
-accommodations.forEach(accommodation => {
-    const name = accommodation.accommodationId
-    if(accommodationArr.hasOwnProperty(name)){
-    accommodationArr[name] = accommodationArr[name] + 1
-    } else{
-        accommodationArr[name] = 1
-    }
-});
-let sortedAccomodatio = []
-for( const key in accommodationArr){
-    sortedAccomodatio.push([key,accommodationArr[key]])
-}
-sortedAccomodatio.sort(function(a, b) {
-    return b[1] - a[1];
-});
-
+    // Get top 5 visited Location 
+    let sortedLocations = await top5VisitedLocation()
+    // Get top 5 visited, Booked Accommodation
+    let sortedAccomodation =await  top5BookedAccommodation()
     res.status(200).send({
         title: 'STATISTICS',
         numberOfActiveUsers: activeUsers.length,
@@ -62,10 +26,8 @@ sortedAccomodatio.sort(function(a, b) {
         numberOfAccommodation: numberOfAccommodation.count,
         numberOfLocation,
         mostVisitedLocations: sortedLocations.slice(0, 5),
-        top5VisitedBookedAccommodations: sortedAccomodatio.slice(0,5)
-    
+        top5VisitedBookedAccommodations: sortedAccomodation.slice(0,5)
     })
 }
-
 module.exports = getStatistics
